add rule to ignore dist directory for stickler
Given that webpack would always generate the final JavaScript file in one line with many other efforts to make the final js file light-weight; it makes the final build fail a lot, if not, all ESLint rules.

So, I thought it best we exclude it for stickler. This way, the students won't have a need to tinker with the configuration.