Update linter's workflow for Ruby on Rails
The official recommendation is to use the latest stable version of the 
linters for new projects. Since this repository is mainly used for new 
projects, it should stick to the recommendation.

By removing the explicit version from the workflow, new projects based 
on this repository will use the linter's latest version by default. This 
helps to:
  1. Remove the discrepancy between the development machine using a 
newer version and the workflow using an older version.
  2. Avoid the need to constantly updating this file to change the 
linter's version.
  3. Stick to the recommendation by using the latest stable version by 
default for new projects.